Segregate rewriting
