+++ lib/deep_cover/node/base.rb
@@ -7,10 +7,9 @@ module DeepCover
     include HasChildHandler
     include CanAugmentChildren
     include HasLocal
+    include Rewriting
     extend CheckCompletion
     include FlowAccounting
-
-    has_child_handler('rewrite_%{name}')
 
     attr_reader :index, :parent, :children, :base_node
 
@@ -36,26 +35,6 @@ module DeepCover
 
     def [](v)
       children[v]
-    end
-
-    # Code to add before and after the node for covering purposes
-    def rewrite
-      '%{node}'
-    end
-
-    def resolve_rewrite(rule, context)
-      rule ||= '%{node}'
-      sources = context.tracker_sources
-      rule.split('%{node}').map{|s| s % {local: local_source, **sources} }
-    end
-
-    def rewrite_prefix_suffix
-      parent_prefix, parent_suffix = resolve_rewrite(parent.rewrite_child(self), parent)
-      prefix, suffix = resolve_rewrite(rewrite, self)
-      [
-        "#{parent_prefix}#{prefix}",
-        "#{suffix}#{parent_suffix}"
-      ]
     end
 
     ### Public API
